Added PrintUtils.kt to color output.
This sadly does not work in all terminals (ANSI Color codes need to be enabled/supported). If someone knows how to detect whether the current console supports it, add a check :)

On Windows it works when using Windows Terminal with Powershell. If you are just using powershell, use the tutorbotWithoutWT.ps1 script instead of tutorbot.ps1, this opens a new shell with enabled color.